Обновлён компонент prepod.tests

+++ app/components/admin.prepods/admin.prepods.model.php
@@ -14,14 +14,14 @@ class AdminPrepodsModel extends \UTest\Kernel\Component\Model
     {
         $users = [];
 
-        if ($this->isActionRequest('del_all')) {
+        if ($this->isActionRequest('del_all') && $this->isNativeActionMethod()) {
             foreach ($this->_POST['i'] as $id) {
                 User::user()->delete($id);
             }
             Site::redirect(Site::getUrl());
         }
 
-        if ($this->isActionRequest('newpass_all')) {
+        if ($this->isActionRequest('newpass_all') && $this->isNativeActionMethod()) {
             foreach ($this->_POST['i'] as $id) {
                 if (!intval($id) || $id == User::ADMIN_ID) {
                     continue;
